Vim: Fix fugitive plugin activation
When plugins are piped (installation combined in one line) in vim-plug
and the first plugin has condition, the second plugin activates also
only on that condition.

In this case it resulted that not only Gitv was started on `:Gitv`
command, but also vim-fugitive. Instead vim-fugitive should always be
started for things like syntas highlighting on commit message.